Clean up
git-svn-id: file:///home/bacon/SVN/spcm/cluster-admin-0.3@274 fc83dafb-f549-e211-8a70-0017315c2eee

+++ install.sh
@@ -48,7 +48,9 @@ cp Common/*.awk ${DESTDIR}${PREFIX}/libexec
 
 mkdir -p ${DESTDIR}${DATADIR}/WWW
 cp Common/Share/* ${DESTDIR}${DATADIR}
-cp $os/Share/* ${DESTDIR}${DATADIR}
+if [ -e $os/Share ]; then
+    cp $os/Share/* ${DESTDIR}${DATADIR}
+fi
 cp $os/WWW/* ${DESTDIR}${DATADIR}/WWW
 
 sed -e "s|add-gecos.awk|${PREFIX}/libexec/add-gecos.awk|g" \
